FE-51 Refactor : 에러수정 1. pages 수정시 발생한 에러를 해결했습니다.
