+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
-import ChatList from './pages/ChatList';
-import ChatRoom from './pages/ChatRoom';
+import ChatList from './pages/ChatList/ChatList';
+import ChatRoom from './pages/Chatroom/ChatRoom';
 import EmailLogin from './pages/EmailLogin/EmailLogin';
 import EmailSignUp from './pages/EmailSignUp/EmailSignUp';
 import Home from './pages/Home/Home';
